requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
blis==1.0.2
boto3==1.40.32
botocore==1.40.32
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
cymem==2.0.11
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.8
huggingface-hub==0.34.4
HyperPyYAML==1.2.2
idna==3.10
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
julius==0.2.7
kiwisolver==1.4.9
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.3
lightning-utilities==0.15.2
llvmlite==0.44.0
Mako==1.3.10
marisa-trie==1.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
primePy==1.3
propcache==0.3.2
protobuf==4.25.8
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pyfhel==3.5.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.3
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
SQLAlchemy==2.0.43
srsly==2.5.1
starlette==0.47.2
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
thinc==9.1.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.19.1
tomli==2.2.1
torch==2.8.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.8.0
torchmetrics==1.8.1
tqdm==4.67.1
transformers==4.44.0
triton==3.4.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
weasel==0.4.1
webrtcvad==2.0.10
wrapt==1.17.3
yarl==1.20.1
